+++ port/lib/src/clipboard.service.dart
@@ -1,0 +1,6 @@
+import 'models/clipboard_data.dart';
+
+abstract class Clipboard {
+  Future<ClipboardData> getData();
+  Future<void> setData(ClipboardData data);
+}