+++ port/lib/src/commands/get_data_from_clipboard.command.dart
@@ -1,0 +1,12 @@
+import '_.imports.dart';
+
+class GetDataFromClipboardCommand implements AsyncCommand<ClipboardData> {
+  final Clipboard clipboard;
+
+  const GetDataFromClipboardCommand(this.clipboard);
+
+  @override
+  Future<ClipboardData> execute() {
+    return clipboard.getData();
+  }
+}