+++ port/lib/src/commands/get_text_from_clipboard.command.dart
@@ -1,0 +1,13 @@
+import '_.imports.dart';
+
+class GetTextFromClipboardCommand implements AsyncCommand<String> {
+  final Clipboard clipboard;
+
+  const GetTextFromClipboardCommand(this.clipboard);
+
+  @override
+  Future<String> execute() async {
+    final data = await clipboard.getData();
+    return data.text;
+  }
+}