+++ port/lib/src/commands/set_data_to_clipboard.command.dart
@@ -1,0 +1,14 @@
+import '_.imports.dart';
+
+class SetDataToClipboardCommand implements AsyncCommand<ClipboardData> {
+  final ClipboardData data;
+  final Clipboard clipboard;
+
+  const SetDataToClipboardCommand(this.clipboard, this.data);
+
+  @override
+  Future<ClipboardData> execute() async {
+    await clipboard.setData(data);
+    return data;
+  }
+}