+++ port/lib/src/commands/set_text_to_clipboard.command.dart
@@ -1,0 +1,15 @@
+import '_.imports.dart';
+
+class SetTextToClipboardCommand implements AsyncCommand<ClipboardData> {
+  final String text;
+  final Clipboard clipboard;
+
+  const SetTextToClipboardCommand(this.clipboard, this.text);
+
+  @override
+  Future<ClipboardData> execute() async {
+    final data = ClipboardData(text: text);
+    await clipboard.setData(data);
+    return data;
+  }
+}