+++ port/lib/src/models/_.imports.dart
@@ -1,0 +1,2 @@
+export 'package:zam_core/equatable.dart' show EquatableMixin;
+export 'package:zam_core/meta.dart' show immutable;