+++ port/test/_setup.dart
@@ -1,0 +1,13 @@
+import '../lib/zam_clipboard.dart';
+
+class MockClipboard extends Clipboard {
+  ClipboardData _data = ClipboardData.initialized();
+
+  Future<ClipboardData> getData() async {
+    return _data;
+  }
+
+  Future<void> setData(ClipboardData data) async {
+    _data = data;
+  }
+}