+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # 0.0.1
 
-- Includes @component_1@.
-- Includes @component_2@.
-- Includes @component_3@.
+- Includes `Clipboard`
+- Includes `ClipboardData`
+- Includes `GetDataFromClipboardCommand`
+- Includes `SetDataToClipboardCommand`
+- Includes `GetTextFromClipboardCommand`
+- Includes `SetTextToClipboardCommand`